pyproject.toml
[project]
name = "telegram-rag-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "jq>=1.10.0",
    "langchain>=1.0.4",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "sentence-transformers>=5.1.2",
]
